Add some basic console interaction
Allow writing things, and having those things show up in the log when
you press enter.
Right now, any inputs that are not LF, keypad enter, or backspace are
just inserted directly into the current command. I think this is
probably problematic (input is of type int, and we're sticking it in a
spot for a char), but it works so far...

Also, rearrange the format of the cmdlog array because it was entirely
wrong (number of strings * size, not size of strings * number, whoops!)

Remove some debugging code too.